Add custom scp options
With the release of macOS Venture the version of `scp` binary used by Belt to upload various assets such as toolbox, archives etc. has fundamentally changed. In Venture scp now requires SFTP support from the host to which we are uploading to.

This PR introduces a variable name `BELT_SCP_OPTIONS` that allows Mac users to switch to using legacy version of `scp`. This can be done by adding `BELT_SCP_OPTIONS="-O"` at the top of their deployment script.